[#1793] Use counter of type long

+++ repository-core/src/main/java/nl/knaw/huygens/repository/storage/mongo/MongoStorageBase.java
@@ -59,7 +59,7 @@ public class MongoStorageBase {
   public static class Counter {
     @JsonProperty("_id")
     public String id;
-    public int next;
+    public long next;
   }
 
 }
